--- tests/index_code_rel_ths.json
@@ -1,1 +1,0 @@
-{"000819": "1B0819", "1B0819": "000819", "000006": "1B0004", "1B0004": "000006", "000823": "1B0823", "1B0823": "000823", "000867": "1B0867", "1B0867": "000867", "000989": "1B0989", "1B0989": "000989", "000987": "1B0987", "1B0987": "000987", "000033": "1B0033", "1B0033": "000033", "000071": "1B0071", "1B0071": "000071", "000035": "1B0035", "1B0035": "000035", "000854": "1B0854", "1B0854": "000854", "000107": "1B0107", "1B0107": "000107", "000092": "1B0092", "1B0092": "000092", "000126": "1B0126", "1B0126": "000126", "000026": "1B0026", "1B0026": "000026", "000145": "1B0145", "1B0145": "000145", "000073": "1B0073", "1B0073": "000073", "000094": "1B0094", "1B0094": "000094", "000105": "1B0105", "1B0105": "000105", "000869": "1B0869", "1B0869": "000869", "000027": "1B0027", "1B0027": "000027", "000135": "1B0135", "1B0135": "000135", "000066": "1B0066", "1B0066": "000066", "000068": "1B0068", "1B0068": "000068", "000692": "1B0692", "1B0692": "000692", "000932": "1B0932", "1B0932": "000932", "000170": "1B0170", "1B0170": "000170", "000036": "1B0036", "1B0036": "000036", "000903": "1B0903", "1B0903": "000903", "000057": "1B0057", "1B0057": "000057", "000016": "1B0016", "1B0016": "000016", "000063": "1B0063", "1B0063": "000063", "000074": "1B0074", "1B0074": "000074", "000055": "1B0055", "1B0055": "000055", "000052": "1B0052", "1B0052": "000052", "000914": "1B0914", "1B0914": "000914", "000028": "1B0028", "1B0028": "000028", "000069": "1B0069", "1B0069": "000069", "000043": "1B0043", "1B0043": "000043", "000102": "1B0102", "1B0102": "000102", "000992": "1B0992", "1B0992": "000992", "000934": "1B0934", "1B0934": "000934", "000849": "1B0849", "1B0849": "000849", "000689": "1B0689", "1B0689": "000689", "000048": "1B0048", "1B0048": "000048", "000827": "1B0827", "1B0827": "000827", "000300": "1B0300", "1B0300": "000300", "000021": "1B0021", "1B0021": "000021", "000123": "1B0123", "1B0123": "000123", "000038": "1B0038", "1B0038": "000038", "000093": "1B0093", "1B0093": "000093", "000853": "1B0853", "1B0853": "000853", "000030": "1B0030", "1B0030": "000030", "000059": "1B0059", "1B0059": "000059", "000117": "1B0117", "1B0117": "000117", "000155": "1B0155", "1B0155": "000155", "000906": "1B0906", "1B0906": "000906", "000010": "1B0007", "1B0007": "000010", "000049": "1B0049", "1B0049": "000049", "000050": "1B0050", "1B0050": "000050", "000974": "1B0974", "1B0974": "000974", "000104": "1B0104", "1B0104": "000104", "000125": "1B0125", "1B0125": "000125", "000141": "1B0141", "1B0141": "000141", "000018": "1B0018", "1B0018": "000018", "000019": "1B0019", "1B0019": "000019", "000003": "1A0003", "1A0003": "000003", "000029": "1B0029", "1B0029": "000029", "000054": "1B0054", "1B0054": "000054", "000096": "1B0096", "1B0096": "000096", "000697": "1B0697", "1B0697": "000697", "000031": "1B0031", "1B0031": "000031", "000137": "1B0137", "1B0137": "000137", "000119": "1B0119", "1B0119": "000119", "000149": "1B0149", "1B0149": "000149", "000056": "1B0056", "1B0056": "000056", "000047": "1B0047", "1B0047": "000047", "000051": "1B0051", "1B0051": "000051", "000108": "1B0108", "1B0108": "000108", "000053": "1B0053", "1B0053": "000053", "000058": "1B0058", "1B0058": "000058", "000901": "1B0901", "1B0901": "000901", "000986": "1B0986", "1B0986": "000986", "000159": "1B0159", "1B0159": "000159", "000133": "1B0133", "1B0133": "000133", "000132": "1B0132", "1B0132": "000132", "000158": "1B0158", "1B0158": "000158", "950096": "1B0865", "1B0865": "950096", "000090": "1B0090", "1B0090": "000090", "000928": "1B0928", "1B0928": "000928", "000032": "1B0032", "1B0032": "000032", "000147": "1B0147", "1B0147": "000147", "000148": "1B0148", "1B0148": "000148", "000905": "1B0905", "1B0905": "000905", "000064": "1B0064", "1B0064": "000064", "000091": "1B0091", "1B0091": "000091", "000011": "1B0008", "1B0008": "000011", "000687": "1B0687", "1B0687": "000687", "000060": "1B0060", "1B0060": "000060", "000004": "1B0001", "1B0001": "000004", "000856": "1B0856", "1B0856": "000856", "000065": "1B0065", "1B0065": "000065", "000070": "1B0070", "1B0070": "000070", "000098": "1B0098", "1B0098": "000098", "000847": "1B0847", "1B0847": "000847", "000100": "1B0100", "1B0100": "000100", "991001": "1C0003", "1C0003": "991001", "000067": "1B0067", "1B0067": "000067", "000852": "1B0852", "1B0852": "000852", "000005": "1B0002", "1B0002": "000005", "000802": "1B0802", "1B0802": "000802", "000020": "1B0020", "1B0020": "000020", "000001": "1A0001", "1A0001": "000001", "000017": "1B0017", "1B0017": "000017", "000134": "1B0134", "1B0134": "000134", "000002": "1A0002", "1A0002": "000002", "000099": "1B0099", "1B0099": "000099", "000076": "1B0076", "1B0076": "000076", "000044": "1B0044", "1B0044": "000044", "000982": "1B0982", "1B0982": "000982", "000009": "1B0009", "1B0009": "000009", "000046": "1B0046", "1B0046": "000046", "000103": "1B0103", "1B0103": "000103", "000891": "1B0891", "1B0891": "000891", "000095": "1B0095", "1B0095": "000095", "000160": "1B0160", "1B0160": "000160", "000851": "1B0851", "1B0851": "000851", "000690": "1B0690", "1B0690": "000690", "000008": "1B0006", "1B0006": "000008", "000045": "1B0045", "1B0045": "000045", "000161": "1B0161", "1B0161": "000161", "000153": "1B0153", "1B0153": "000153", "000146": "1B0146", "1B0146": "000146", "000913": "1B0913", "1B0913": "000913", "000139": "1B0139", "1B0139": "000139", "000042": "1B0042", "1B0042": "000042", "000129": "1B0129", "1B0129": "000129", "000122": "1B0122", "1B0122": "000122", "000062": "1B0062", "1B0062": "000062", "000693": "1B0693", "1B0693": "000693", "000115": "1B0115", "1B0115": "000115", "000106": "1B0106", "1B0106": "000106", "000034": "1B0034", "1B0034": "000034", "000040": "1B0040", "1B0040": "000040", "000110": "1B0110", "1B0110": "000110", "000128": "1B0128", "1B0128": "000128", "000072": "1B0072", "1B0072": "000072", "000118": "1B0118", "1B0118": "000118", "000860": "1B0860", "1B0860": "000860", "000120": "1B0120", "1B0120": "000120", "000991": "1B0991", "1B0991": "000991", "000114": "1B0114", "1B0114": "000114", "000933": "1B0933", "1B0933": "000933", "000142": "1B0142", "1B0142": "000142", "000152": "1B0152", "1B0152": "000152", "000111": "1B0111", "1B0111": "000111", "000097": "1B0097", "1B0097": "000097", "000814": "1B0814", "1B0814": "000814", "000138": "1B0138", "1B0138": "000138", "000695": "1B0695", "1B0695": "000695", "000136": "1B0136", "1B0136": "000136", "000037": "1B0037", "1B0037": "000037", "000858": "1B0858", "1B0858": "000858", "000993": "1B0993", "1B0993": "000993", "000935": "1B0935", "1B0935": "000935", "000688": "1B0688", "1B0688": "000688", "000078": "1B0078", "1B0078": "000078", "000162": "1B0162", "1B0162": "000162", "000015": "1B0015", "1B0015": "000015", "000116": "1B0116", "1B0116": "000116", "000061": "1B0061", "1B0061": "000061", "000863": "1B0863", "1B0863": "000863", "000022": "1B0022", "1B0022": "000022", "000101": "1B0101", "1B0101": "000101", "000012": "1B0012", "1B0012": "000012", "000683": "1B0683", "1B0683": "000683", "000007": "1B0005", "1B0005": "000007", "000130": "1B0130", "1B0130": "000130", "000079": "1B0079", "1B0079": "000079", "000077": "1B0077", "1B0077": "000077", "000113": "1B0113", "1B0113": "000113", "000075": "1B0075", "1B0075": "000075", "000121": "1B0121", "1B0121": "000121", "000857": "1B0857", "1B0857": "000857", "000682": "1B0682", "1B0682": "000682", "000025": "1B0025", "1B0025": "000025", "000041": "1B0041", "1B0041": "000041", "000039": "1B0039", "1B0039": "000039", "000151": "1B0151", "1B0151": "000151", "000109": "1B0109", "1B0109": "000109", "000150": "1B0150", "1B0150": "000150", "000112": "1B0112", "1B0112": "000112", "000685": "1B0685", "1B0685": "000685", "000131": "1B0131", "1B0131": "000131"}
